+++ gridstatus/tests/test_ercot.py
@@ -23,3 +23,28 @@ def test_ercot_get_fuel_mix():
     df = iso.get_fuel_mix("latest").mix
     assert df.shape[0] >= 0
     assert df.columns.tolist() == ["Time", "Solar", "Wind", "Other"]
+
+
+def test_ercot_get_as_prices():
+    as_cols = [
+        "Time",
+        "Market",
+        "Non-Spinning Reserves",
+        "Regulation Down",
+        "Regulation Up",
+        "Responsive Reserves",
+    ]
+
+    # today
+    iso = gridstatus.Ercot()
+    today = pd.Timestamp.now().date()
+    df = iso.get_as_prices(today)
+    assert df.shape[0] >= 0
+    assert df.columns.tolist() == as_cols
+    assert df["Time"].unique()[0].date() == today
+
+    date = today - pd.Timedelta(days=3)
+    df = iso.get_as_prices(date)
+    assert df.shape[0] >= 0
+    assert df.columns.tolist() == as_cols
+    assert df["Time"].unique()[0].date() == date