+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add ability to save intermediate results to disk when fetching data across multiple requests using `save_to` parameter
 - Add ability to get renewable curtailment data for CAISO
 - Add Ancillary Service methods for CAISO
+- Add Ancillary Service Prices for ERCOT
 
 ## v0.14.0 - Nov 8, 2022